--- a/test/Exceptionless.Performance.Tests/Program.cs
+++ b/test/Exceptionless.Performance.Tests/Program.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Running;
+
+namespace Exceptionless.Performance.Tests
+{
+    class Program
+    {
+        static void Main(string[] args) {
+            var summaries = BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly)
+                .Run(args);
+        }
+    }
+}