Up to x3.5 faster snake casing for Json properties via caching; also dumps the cache at Gen2.

Move to `ConversionCached` approach; less allocations and much faster conversion (also incorporates the `StringBuilderCached` for new strings)

|              Method |           Word |      Mean | Ratio |  Gen 0 | Allocated |
|-------------------- |--------------- |----------:|------:|-------:|----------:|
|           SnakeCase |              A |  37.22 ns |  1.00 | 0.0255 |     120 B |
| StringBuilderCached |              A |  35.24 ns |  0.94 | 0.0051 |      24 B |
|    ConversionCached |              A |  18.99 ns |  0.51 |      - |         - |
|                     |                |           |       |        |           |
|           SnakeCase | InnerException | 109.54 ns |  1.00 | 0.0374 |     176 B |
| StringBuilderCached | InnerException | 101.12 ns |  0.92 | 0.0118 |      56 B |
|    ConversionCached | InnerException |  23.95 ns |  0.22 |      - |         - |
|                     |                |           |       |        |           |
|           SnakeCase |        Message |  65.09 ns |  1.00 | 0.0322 |     152 B |
| StringBuilderCached |        Message |  59.97 ns |  0.92 | 0.0085 |      40 B |
|    ConversionCached |        Message |  22.86 ns |  0.35 |      - |         - |
|                     |                |           |       |        |           |
|           SnakeCase |     StackTrace |  91.76 ns |  1.00 | 0.0340 |     160 B |
| StringBuilderCached |     StackTrace |  82.57 ns |  0.90 | 0.0101 |      48 B |
|    ConversionCached |     StackTrace |  22.07 ns |  0.24 |      - |         - |
|                     |                |           |       |        |           |
|           SnakeCase |       property |  55.52 ns |  1.00 | 0.0322 |     152 B |
| StringBuilderCached |       property |  55.52 ns |  1.00 | 0.0085 |      40 B |
|    ConversionCached |       property |  21.86 ns |  0.39 |      - |         - |