@ejsmith 2021-04-19
I guess this is highly likely to hit cache as well since JSON names are repeated so much, eh? What kind of overall difference does this make?