@ejsmith 2021-04-23
@benaadams my concern with this PR is that it's pretty hard core and I'd just like to see that it's worth it outside of the microbenchmark of just snake casing strings. I'm pretty sure this is going to get called a ton in the real world and it would get a lot of cache hits because JSON has a lot of repeated strings, but it would be really nice to see an overall benchmark of using the client to send a bunch of events and seeing how changes like this and hopefully many more in the future impact the client overall. Thoughts?