@benaadams 2021-05-10
Saying that; there's a bigger item I also want to add and it will use the same GC mechanism, so will add that first and come back with measurements